turn-detection-demo/pyproject.toml
[project]
name = "turn-detection-demo"
version = "0.1.0"
description = "Pipecat voice agent comparing VAD-only vs semantic turn detection over Plivo telephony"
requires-python = ">=3.11"
dependencies = [
    "pipecat-ai[google,deepgram,cartesia,silero,local-smart-turn-v3,runner]",
    "python-dotenv",
    "loguru",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]

[tool.setuptools]
py-modules = []

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
